Go-get example problems fixed.
